cirrus: fix wrong playback rate when switching between 44.1kHz and 48kHz
Because of delayed DAPM powerdown FLL1 can still be running
when another playback is started and the FLL won't be reconfigured
in set_bias_level.

If the current stream requires a different FLL1 rate, for example
when switching between 44.1kHz-related and 48kHz-related rates,
we are running at a wrong rate.

So keep track of the current FLL1 rate and manually change it during
hw_params if needed.